refactor(mobile-markdown): split bridge save guard tests
